some Edits
